Cleaned WindowPattern to its (hopefully) final form
