+++ src/FlaUI.Core/Patterns/IWindowPattern.cs
@@ -6,17 +6,36 @@ namespace FlaUI.Core.Patterns
 {
     public interface IWindowPattern : IPatternWithInformation<IWindowPatternInformation>
     {
+        IWindowPatternProperties Properties { get; }
+        IWindowPatternEvents Events { get; }
+        void Close();
+        void SetWindowVisualState(WindowVisualState state);
+        int WaitForInputIdle(int milliseconds);
+    }
+
+    public interface IWindowPatternProperties
+    {
         PropertyId CanMaximizeProperty { get; }
         PropertyId CanMinimizeProperty { get; }
         PropertyId IsModalProperty { get; }
         PropertyId IsTopmostProperty { get; }
         PropertyId WindowInteractionStateProperty { get; }
         PropertyId WindowVisualStateProperty { get; }
+    }
 
-        void Close();
+    public interface IWindowPatternEvents
+    {
+        EventId WindowClosedEvent { get; }
+        EventId WindowOpenedEvent { get; }
+    }
 
-        void SetWindowVisualState(WindowVisualState state);
-
-        int WaitForInputIdle(int milliseconds);
+    public interface IWindowPatternInformation : IPatternInformation
+    {
+        bool CanMaximize { get; }
+        bool CanMinimize { get; }
+        bool IsModal { get; }
+        bool IsTopmost { get; }
+        WindowInteractionState WindowInteractionState { get; }
+        WindowVisualState WindowVisualState { get; }
     }
 }
